Changed MODE command and added QUIT

+++ srcs/execute/Executer.cpp
@@ -12,6 +12,8 @@ void Execute(Server *server, Client *client)
 	if (client->GetStatus() == DELETED)
 		return ;
 	
+	// To check if the client is still exists after execution.
+	int fd = client->GetSocket();
 	cmd = client->GetCmdBuff();
 	// Iterate through the commands.
 	while (cmd != NULL)
@@ -37,13 +39,18 @@ void Execute(Server *server, Client *client)
 		if (DEBUG)
 			std::cout << "Execute command: " << cmd->command << std::endl;
 
-		// TODO delete after commands realisation.
 		// Trying to find a command and execute it.
 		if (command_funcs.find(cmd->command) != command_funcs.end()
 			&& (cmd->command != "PASS" || client->GetStatus() == PASSWORD))
 			command_funcs[cmd->command](server, client, cmd);
 		else if (cmd->command != "CAP")
 			client->_send_buff.append(ERR_UNKNOWN_COMMAND(client->GetNickname(), cmd->command));
+		
+		std::cout << "Here execute" << std::endl;
+		std::map<int, Client*> *clients = server->GetClients();
+		if (clients->find(fd) == clients->end())
+			throw new std::runtime_error("Client disconected");
+		std::cout << "Here execute2" << std::endl;
 		cmdList *next = cmd->next;
 		client->DeleteCommand(cmd);
